# packages/api-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ReCredita API",
    "description": "Recuperação (Plano A) e aquisição (Plano B) de veículos.",
    "version": "0.1.0"
  },
  "components": {
    "schemas": {
...
      "DistribuirInput": {
        "type": "object",
        "properties": {
          "recuperadorId": {
            "type": "string",
            "minLength": 1
          },
          "horasParaAceite": {
            "default": 24,
            "type": "integer",
            "minimum": 1,
            "maximum": 168
          },
          "diasDePrazo": {
            "default": 10,
            "type": "integer",
            "minimum": 1,
            "maximum": 180
          }
        },
        "required": [
          "recuperadorId"
        ],
        "additionalProperties": false
      },
      "RevelarDevedorInput": {
        "type": "object",
        "properties": {
          "finalidade": {
            "type": "string",
            "minLength": 10
          }
        },
        "required": [
          "finalidade"
        ],
        "additionalProperties": false
      },
...
      "Erro": {
        "type": "object",
        "properties": {
          "erro": {
            "type": "string"
          }
        },
        "required": [
          "erro"
        ],
        "additionalProperties": false
      },
...
      "CasoDetalhe": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "ativoId": {
            "type": "string"
          },
          "fonteId": {
            "type": "string"
          },
          "origem": {
            "type": "string",
            "enum": [
              "plataforma_credor",
              "lead_proprio"
            ]
          },
          "finalidade": {
            "type": "string",
            "enum": [
              "recuperacao_para_credor",
              "aquisicao_com_quitacao"
            ]
          },
          "status": {
            "type": "string"
          },
          "recuperadorId": {
            "nullable": true,
            "type": "string"
          },
          "recuperadorNome": {
            "nullable": true,
            "type": "string"
          },
          "prazoVinculo": {
            "nullable": true,
            "type": "string",
            "format": "date-time"
          },
          "prazoMaximo": {
            "nullable": true,
            "type": "string",
            "format": "date-time"
          },
          "canalLead": {
            "nullable": true,
            "type": "string"
          },
          "evidenciaLead": {
            "nullable": true,
            "type": "string"
          },
          "saldoDevedor": {
            "nullable": true,
            "type": "number"
          },
          "valorQuitacaoNegociado": {
            "nullable": true,
            "type": "number"
          },
          "valorPagoAoDevedor": {
            "nullable": true,
            "type": "number"
          },
          "anuenciaCredor": {
            "nullable": true,
            "type": "string",
            "enum": [
              "Pendente",
              "Obtida",
              "Negada"
            ]
          },
          "renajudAtivo": {
            "nullable": true,
            "type": "boolean"
          },
          "gravameBaixado": {
            "nullable": true,
            "type": "boolean"
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "atualizadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "placa": {
            "type": "string"
          },
          "modelo": {
            "nullable": true,
            "type": "string"
          },
          "cidade": {
            "nullable": true,
            "type": "string"
          },
          "uf": {
            "nullable": true,
            "type": "string"
          },
          "ativo": {
            "type": "object",
            "properties": {
              "chassi": {
                "nullable": true,
                "type": "string"
              },
              "ano": {
                "nullable": true,
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "cor": {
                "nullable": true,
                "type": "string"
              },
              "credorNome": {
                "nullable": true,
                "type": "string"
              },
              "valorDivida": {
                "nullable": true,
                "type": "number"
              },
              "dataRecebimento": {
                "type": "string",
                "format": "date-time"
              },
              "renavam": {
                "nullable": true,
                "type": "string"
              },
              "valorFipe": {
                "nullable": true,
                "type": "number"
              },
              "fipeCodigo": {
                "nullable": true,
                "type": "string"
              },
              "fipeReferencia": {
                "nullable": true,
                "type": "string"
              },
              "marca": {
                "nullable": true,
                "type": "string"
              },
              "situacao": {
                "nullable": true,
                "type": "string"
              },
              "tipo": {
                "nullable": true,
                "type": "string"
              },
              "especie": {
                "nullable": true,
                "type": "string"
              },
              "categoria": {
                "nullable": true,
                "type": "string"
              },
              "carroceria": {
                "nullable": true,
                "type": "string"
              },
              "combustivel": {
                "nullable": true,
                "type": "string"
              },
              "potencia": {
                "nullable": true,
                "type": "string"
              },
              "cilindradas": {
                "nullable": true,
                "type": "string"
              },
              "motor": {
                "nullable": true,
                "type": "string"
              },
              "procedencia": {
                "nullable": true,
                "type": "string"
              },
              "municipioEmplacamento": {
                "nullable": true,
                "type": "string"
              },
              "ufEmplacamento": {
                "nullable": true,
                "type": "string"
              },
              "anoFabricacao": {
                "nullable": true,
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "chassi",
              "ano",
              "cor",
              "credorNome",
              "valorDivida",
              "dataRecebimento",
              "renavam",
              "valorFipe",
              "fipeCodigo",
              "fipeReferencia",
              "marca",
              "situacao",
              "tipo",
              "especie",
              "categoria",
              "carroceria",
              "combustivel",
              "potencia",
              "cilindradas",
              "motor",
              "procedencia",
              "municipioEmplacamento",
              "ufEmplacamento",
              "anoFabricacao"
            ],
            "additionalProperties": false
          },
          "fonte": {
            "type": "object",
            "properties": {
              "nome": {
                "type": "string"
              },
              "tipo": {
                "type": "string",
                "enum": [
                  "Plataforma",
                  "Credor Direto",
                  "Lead Próprio"
                ]
              },
              "credorNome": {
                "nullable": true,
                "type": "string"
              }
            },
            "required": [
              "nome",
              "tipo",
              "credorNome"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "id",
          "ativoId",
          "fonteId",
          "origem",
          "finalidade",
          "status",
          "recuperadorId",
          "recuperadorNome",
          "prazoVinculo",
          "prazoMaximo",
          "canalLead",
          "evidenciaLead",
          "saldoDevedor",
          "valorQuitacaoNegociado",
          "valorPagoAoDevedor",
          "anuenciaCredor",
          "renajudAtivo",
          "gravameBaixado",
          "criadoEm",
          "atualizadoEm",
          "placa",
          "modelo",
          "cidade",
          "uf",
          "ativo",
          "fonte"
        ],
        "additionalProperties": false
      },
      "EventoCaso": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "tipo": {
            "type": "string",
            "enum": [
              "criado",
              "status_alterado",
              "distribuido",
              "rito_definido",
              "registro_juridico",
              "verificacao_veicular",
              "avistamento",
              "relatorio_colado"
            ]
          },
          "statusDe": {
            "nullable": true,
            "type": "string"
          },
          "statusPara": {
            "nullable": true,
            "type": "string"
          },
          "usuarioNome": {
            "nullable": true,
            "type": "string"
          },
          "ocorridoEm": {
            "type": "string",
            "format": "date-time"
          },
          "dados": {
            "type": "object",
            "additionalProperties": {}
          }
        },
        "required": [
          "id",
          "tipo",
          "statusDe",
          "statusPara",
          "usuarioNome",
          "ocorridoEm",
          "dados"
        ],
        "additionalProperties": false
      },
      "DevedorRevelado": {
        "type": "object",
        "properties": {
          "devedorNome": {
            "nullable": true,
            "type": "string"
          },
          "devedorDoc": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "devedorNome",
          "devedorDoc"
        ],
        "additionalProperties": false
      },
      "AcessoDadoPessoal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "usuarioNome": {
            "type": "string"
          },
          "campos": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "finalidade": {
            "type": "string"
          },
          "acessadoEm": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "usuarioNome",
          "campos",
          "finalidade",
          "acessadoEm"
        ],
        "additionalProperties": false
      },
      "Distribuicao": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "recuperadorId": {
            "type": "string"
          },
          "prazoVinculo": {
            "type": "string",
            "format": "date-time"
          },
          "prazoMaximo": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "status",
          "recuperadorId",
          "prazoVinculo",
          "prazoMaximo"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/api/casos/{id}": {
      "get": {
        "tags": [
          "casos"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CasoDetalhe"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/api/casos/{id}/eventos": {
      "get": {
        "tags": [
          "casos"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventoCaso"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/casos/{id}/devedor": {
      "post": {
        "tags": [
          "casos"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevelarDevedorInput"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevedorRevelado"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/api/casos/{id}/acessos": {
      "get": {
        "tags": [
          "casos"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AcessoDadoPessoal"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/casos/{id}/distribuir": {
      "post": {
        "tags": [
          "casos"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DistribuirInput"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Distribuicao"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
...
  }
}